+++ nets/wide_resnet.py
@@ -56,7 +56,7 @@ class WideResNet(nn.Module):
         super(WideResNet, self).__init__()
         nChannels = [16, 16*widen_factor, 32*widen_factor, 64*widen_factor]
         assert((depth - 4) % 6 == 0)
-        n = (depth - 4) / 6
+        n = int((depth - 4) / 6)
         block = BasicBlock
         # 1st conv before any network block
         self.conv1 = nn.Conv2d(
